backend/requirements.txt
python-dotenv
uvicorn
mysql-connector-python
fastapi
PyJWT
python-multipart
